lib/widgets/statistics/pie_section.dart extract pie component
